requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.2.0
streamlit==1.29.0
openai==0.28.1
python-dotenv==1.0.0
html5lib==1.1
urllib3==2.0.7
numpy==1.26.4 